// app/(audio)/type.ts
export const enum MessageType {
  BROWSER_START_RECORD_AUDIO = "BROWSER_START_RECORD_AUDIO",
  RN_START_RECORD_AUDIO_SUCCESS = "RN_START_RECORD_AUDIO_SUCCESS",
  BROWSER_STOP_RECORD_AUDIO = "BROWSER_STOP_RECORD_AUDIO",
  RN_STOP_RECORD_AUDIO_SUCCESS = "RN_STOP_RECORD_AUDIO_SUCCESS",
  BROWSER_PLAY_AUDIO = "BROWSER_PLAY_AUDIO",
  RN_PLAY_AUDIO_SUCCESS = "RN_PLAY_AUDIO_SUCCESS",
  BROWSER_PAUSE_AUDIO = "BROWSER_PAUSE_AUDIO",
  RN_PAUSE_AUDIO_SUCCESS = "RN_PAUSE_AUDIO_SUCCESS",
}